requirements.txt
asgiref==3.5.2
certifi==2022.12.7
Django==4.1
django-cors-headers==3.13.0
pip==22.3.1
python-decouple==3.7
setuptools==65.6.3
sqlparse==0.4.3
<<<<<<< HEAD
wheel==0.37.1
djangorestframework==3.14.0
=======
djangorestframework==3.14.0
wheel==0.37.1
PyJWT~=2.6.0
>>>>>>> dc11c7c6e52fc6476912a179dbb55bf55670ac0c
